requirements.txt
streamlit
openai
phidata
duckduckgo-search